examples/simple_mlp/requirements.txt
# Requirements for Simple MLP Example

# Core dependencies (required)
hydra-core>=1.3.0
omegaconf>=2.3.0

# Optional ML dependencies (for enhanced functionality)
torch>=1.9.0
numpy>=1.20.0
wandb>=0.15.0

# For installing HSM (if not already installed)
# pip install -e ../../ 